Update the original string
The original document is updated and it needs to be used to generate
the translated version of the document. It is better to update ja.po
to catch up the latest.

Note that "latest" here is HEAD revision of master branch
(commit efe410a887bce532962c689be451e9b6d5875693).

Sync with new operation-guide repository is the future work.