Use n-gram search strategy to create webhelp index
Add <webhelpIndexerLanguage>ja</webhelpIndexerLanguage> to the configuration
section based on the advise from David on openstack-docs ML.

----
If you build a webhelp version, please add the following to the
appropriate <configuration> section of the pom:

<webhelpIndexerLanguage>ja</webhelpIndexerLanguage>

With that set, the search should work using an n-gram search strategy,
taking into account the fact that Japanese doesn't have spaces between
words to use when creating an index. So it looks for any sentence in
which that combination of characters occurs. Not perfect, but probably
the best you can do with a JavaScript-powered client-side search.